+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fixed hooking in ProtocolLib
 - max_npc_distance was set to 5.3 to prevent instant quiting of conversations
 - conversation IO menu sometimes leave an armorstand spawned
+- sometimes messages in a conversation are not send when packet interceptor is used
 ### Security
 
 ## [1.12.4] - 2021-07-05